Add api controller
